Use the new name-finder in a bunch of places
In most cases, the meaning is the same, but we are now much smarter
about what is a string.  This means we temporarily do worse for mocks,
but I have plans for those.  We handle better a number of cases where
we'd previously think there might be an implicit import.  And we fix a
bunch of bugs in the conflicts-check, which apparently I had never
tested.