Big rewrite to remove khodemod and use ASTTokens
This is a big rewrite of many parts of slicker.  The main purpose is to
rip out `codemod` and replace it with a custom thing.  Hopefully
eventually that custom thing will have most of the features of codemod,
but for now it just has the ones we need.  In any case, this simplifies
things in several places, and allows us to be smarter about using the
AST because things can't change under us and we can process on a
character level.  To do this, I had to get more info from the AST, for
which I used a library for mapping AST nodes to tokens.