Fix up string/comment references
We are now much, much smarter; this should do all the things we need for
mocks and other string-references used by code, and do as good a job as
reasonable on human-readable comments and docstrings.  (It won't catch
people who refer to something by neither its fully-qualified name nor
a name by which it is imported in this file, but I think that's just not
something we can really even try to do.  We also aren't very smart about
following linebroken comments, although we could be smarter there.)

The tricky part is that for strings we need to both handle multi-token
strings (`"foo." "bar"`) and strings with escapes in them (we only
handle when the escapes are not in the name itself.)  This requires a
bunch of annoying token-walking and special-casing.