Add a testcase for some previously-untested code.
Summary:
We had a bunch of tests around implicit imports, but apparently not
around renaming a symbol that was only provided implicitly.  Now we
do.  And it turns out that code had a bug!, which is now fixed.

While debugging, I made one small change that made it much easier to
examine diffs around the expected file output.

Test Plan: make test

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D40252