Add a `--root` flag to specify the root of the tree to make changes in.
Summary:
Before, we assumed that you wanted to make changes in every file below
you in the current working directory.  But it would be nice to be able
to run slicker somewhere else, and say "make changes to everything
below directory X."  This adds support for that.

We now have a change in semantics: whereas before all filenames were
taken to be relative to the current working directory, now they're
relative to the value of `--root`.

slicker.py changes surprisingly little; the rewriting code deals with
modules, not paths, and doesn't care what the root is.  khodemod.py
has to pass around the root a lot though since read_file() and
write_file() obviously care.

Test Plan: make test

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D40258