Get rid of the TestFrontend, and make read_file() global.
Summary:
We use temp dirs instead of TestFrontend everywhere, with mocking to
test the warning and error output.

This lets us move read_file() to be a global function, which is
necessary for some of the symbol-moving functionality we want to
write.

I also had to update the fix-includes substate again because somehow
it got reverted at some point.

Test Plan: make test

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D40371