Fill in the symbol move suggestor
Summary:
This only does the easy part -- moving the right region of the file --
so far; we'll need to fix up imports, references to the old file in the
moved region, and references to the moved symbol in the new and old
files.  But anyway, it's a start.  To handle comments correctly I also
fixed a bug in `_get_area_for_ast_node`, where we need to be okay with
DEDENT tokens that are empty; surprisingly `''.isspace()` is `False`.

I also fixed up a line-indexing bug in error handling, which required
fixing up some tests that were asserting an incorrect set of warnings.
And I turned off automove some tests which didn't have things set up for
it and were now failing rather than doing nothing.

(I paired with @csilvers on the core of the new suggestor code.)

Test Plan: make test

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40383