Fix up references to the moved symbol in the old/new files
Summary:
This the first two parts of fixing up references in the old/new files
when moving a symbol.  Here, we fix up the references to the moved
symbol itself.  For the most part, this just means thinking of things as
if the file where a symbol lives has a localname for it which doesn't
correspond to an import; the rest of the code mostly Just Works.

The complicated bit is that we want to add imports differently in this
case: we want to find a reasonable place at the toplevel to add the
import and add it, not copying any context.  For a "reasonable place" we
use the last toplevel import if there is one, and right after any prefix
consisting of comments, docstrings, and newlines if not.

Test Plan: make test

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40390