Use the contextless-import code if there is no explicit import
Summary:
Previously, if there was no explicit import of the old symbol, but there
was an implicit import, we'd try to match that import.  We didn't do it
quite correctly in all cases, and it's not clear it's really worth
trying.  Now, instead, we simply act as if there were no relevant
import, and add the new import somewhere reasonable at the toplevel.  In
most cases the result will be the same.  In edge cases where this adds
circular imports, the user can always fix it up some better way.

Test Plan: make test

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40401