Move all the moving stuff into a separate module
Summary:
It's gotten a bit longer, and will be getting even longer, and is fairly
separate from the rest.  This also meant I had to create a util.py for
shared things.  I did this entire move with `slicker` itself!  It more
or less just worked, with the exception of the known TODO that I need to
fix up imports, and trying to operate on fix_python_imports since the
way I have that included it's not obviously third-party.

Test Plan: make test

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40407