Fix up references inside the moved symbol itself
Summary:
When we move the definition of a symbol, it may reference other things
in the source and/or destination modules as well as itself.  We need to
fix up those references.  This commit does that.

It's still not the cleanest: I'd like to be able to share a lot more
code with `_fix_uses_suggestor`.  But I broke out the biggest and
clearest shared parts, so there's not too much literal duplication.
This primarily involved adding more options to a few helpers to give
more control in what they process.

I also fixed a few issues along the way because I happened to write test
cases that hit them:
- We now strip leading newlines from the moved region, because otherwise
  when moving to a blank file  we'd get a giant space between the newly
  added imports (if any) and the added definitions.  This more or less
  solves the extra-spaces problem.
- I realized there's a better way to handle the issue where we are
  moving `foo.myfunc` to `bar.myfunc` and a file has `import foo as bar`
  so we need to update the import but not any inline references; that's
  the changes to `_replace_in_file`.  (We need to handle this better
  because it includes the case where the pre-move file imported itself.)
  But when writing the tests, I noticed that actually most such cases
  will be spuriously flagged as conflicts; I just left a TODO for that.
- I had the tests assert about expected errors first, since an added
  error is often easier to understand than a missing change to the file.

Note that this doesn't handle moving/duplicating imports of unrelated
modules that are used in the moved region; I'll get to those next.
(That also include removing imports of the new module in the old, if the
only references to those were in the moved region itself.  We already
conversely add an import of the old module in the new, if necessary.)

Test Plan:
- `make lint test`
- `slicker.py slicker._fix_moved_region_suggestor fix_uses.fix_moved_region_suggestor`
  and then do some spot checking and run flake8 and see that the only
  things that it complains about are missing and unused imports and
  whitespace.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40448