Add support for filenames as inputs to slicker, and add a bunch of tests.
Summary:
The tests turned up some problems with the existing code too.
Overall, this was a surprisingly invasive change.  But I'm confident
the common cases are handled due to the new tests.

Make sure we check that the input module already exists.

By default, we were doing that in the case where the input was
specified as a package or module (that is, using dots) because we used
the filesystem to decide what the type was.  But when the input was
specified as a filename, we just blindly accepted it before.  Now we
check.

Just to be extra-safe, we do the checking in both cases -- whether the
input is a filename or a module-name -- even though it will always
pass in the second case.

I also made a few more non-contentful changes based on review.

Test Plan:
make test

make test

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D40440