Fix up imports used in the moved region
Summary:
This is the last major part of making symbol-moves work without any
manual editing: we need to move or copy any imports that are referenced
by the moved region of code.  Like updating the corresponding
references, this is split into two parts, although in this case we find
it more convenient to do the import-adding along with the
reference-fixing.

This is accomplished by two major changes.  First, we modify
`_fix_moved_region_suggestor` to also fix up references to any imported
symbols which changed names, and to track the imports we need to add for
imported symbols.  To do this, we take advantage of some of the
new shared code added in recent refactors -- and with these changes we
end up practically rewriting `_fix_moved_region_suggestor`.

Additionally, we add a new suggestor that cleans up removed imports.
(Actually, we have two: one for the old file and one for the moved
region.  Eventually I'd like to reunify them all.) This proceeds
somewhat similarly to `_fix_imports_suggestor`, although it's much more
aggressive about removing unused imports that are potentially unrelated
to our change, because it's hard to tell by this point what we've
touched, and in any case our changes to this file are already more
invasive.  (This is the source of the few test that aren't removing
TODOs.) For imports in the moved region, we modify `_unused_imports` to
be able to tell it to look at a specific AST node -- in this case the
moved region.

Depends on D40532, D40538, and D40558.

After this, I'll do some splitting up of slicker.py since it's gotten
quite large and slicker itself should be able to do that!

Test Plan:
- `make lint test`
- Ran
  ```
  ./slicker.py slicker._fix_uses_suggestor fix_uses.fix_uses_suggestor
  ./slicker.py slicker._fix_moved_region_suggestor fix_moved_region.fix_moved_region_suggestor
  ```
  and check that `make test` still passed and that the new file looked
  reasonable.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40516