Add suggestors to remove empty files and leading whitespace
Summary:
We do these as after-the-fact cleanups a la import sorting so we don't
have to do a bunch of complicated logic to figure out when removing
imports if we've removed all of them, or what is left behind.  When we
do it that way, it's quite simple!

Test Plan: `make lint test`

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40517