Break out some pieces of _fix_uses_suggestor
Summary:
This is a refactor (broken out of D40516) to break out a few pieces of
functionality in `_fix_uses_suggestor` which are or soon will be used in
`_fix_moved_region_suggestor` as well.  They are:
- `_choose_best_localname` takes care of deciding what imports of
  `fullname` already exist in a file, and therefore what `localname` we
  should use for it there.
- `_import_stmt` computes the correct import statement for a name/alias.
- `_remove_import_patch` returns a patch (or warning) that actually
  removes the given import.

I also changed `_add_contextless_import_patch` to accept imports that
already have the trailing newline.

Test Plan: `make lint test`

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40532