Split import-updating from reference-updating
Summary:
This refactor splits the import-updating part of `_fix_uses_suggestor`
out into `_fix_imports_suggestor`.  In order to do this, we modify
`_imports_to_remove` to simply check for all unused imports, and remove
them, so that it doesn't need to know much about what patches we made,
since we no longer have easy access to that information.  This
allows us to separate concerns much better -- and should be an even
bigger improvement when we get to fixing up imports in the moved region.
The only cost is that we now wantonly remove imports we think are
unused (unless they have an `@UnusedImport`) without checking that
we touched something that we got from them.  But we are still careful to
not touch totally unrelated imports, so this shouldn't be a huge problem
in practice.

Test Plan: `make test lint`

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40538