Move import-adding back along with symbol updating
Summary:
In D40538, I split import-fixing out from reference-fixing, because
import removal is much simpler if we can assume that the changes from
reference-fixing have already been made.  But when reworking D40516, I
realized that this really only applies to import removal, and that
separating out import addition hadn't gotten us anything and in fact had
resulted in some added complexity.  (And it would have been even more
complex to do it separately for the moved region version.)  So now I
move that part back, for consistency and simplicity.

The only functional changes are subtle: first, in certain edge cases
(such as the one in `test_uses_old_module_via_implicit_self_import`)
where we copy an implicit import we now warn about it, because we end up
considering whether to remove it at a point at which we formerly didn't.
This is a super edge case and arguably the new behavior is better
anyway.  Second, in some cases we refer to different lines in warnings,
because we do additions before removals now.  (Neither is necessarily
correct by the time import-sorting has run, so so be it.)

Test Plan: `make lint test`

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40562