Add more terms and such to top-of-file docstrings
Summary:
I realized we didn't have a toplevel place where we defined cases like
"implicit import" and "suggestor" that we talk about all over the place.
Now there is a "special cases" section in slicker.py which deals with
the former (as well as other special cases we handle) and a "terms" list
in khodemod.py which deals with the latter.  I then fixed up inline
comments to mention those, and fixed a few errors in them and updated
and expanded the `make_fixes` docstring while I was at it.

Test Plan: `make lint test`, read the docs

Reviewers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40584