Massively expand the README
Summary:
Hopefully now this should be a useful starting point for anyone who
wants to use slicker.

Depends on D40516 because it claims we do things that are implemented
there.

Test Plan:
read it; push to a branch on github and read it there to test the
markdown renders reasonably

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40587