Compute AST and ASTTokens lazily
Summary:
Computing the `asttokens.ASTTokens` object (the ast -> token mapping)
for a file is surprisingly slow: in my tests it took ~30x as long as
`ast.parse` (which already dwarfs the actual file-read), making it by
far the slowest step.  Now we compute it (and the AST, for good measure)
lazily when we need it.  To make this useful, `Import.start` and `.end`,
which we need the tokens to compute, are also now lazy.  (And to do
that, I made it a proper class rather than a namedtuple.)

Additionally, I did some refactoring such that we only look at tokens
for string-replacement if we're actually doing any replacing; the way I
did it we rerun `get_tokens` a bit more but that's ok because it's super
fast.  I also did the same thing for comments, this time in a somewhat
hackier way because they don't appear in the AST at all -- so I just
looked at the entire body for the regexes.  With those changes, we only
create the `ASTTokens` object for files we actually modify.

Test Plan:
- `make lint test`
- ran slicker over webapp, saw the doing-the-replacements step took ~20s
  rather than ~80s (per replacement)

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40681