Cache path-resolution across runs
Summary:
Resolving which paths to operate on is sadly pretty slow on large
codebases (e.g. with lots of ignored files) and is fairly easy to cache.
So now we do!  We clear the cache if you write a new file, which luckily
you usually only do in the very first automove, which is before the
first time you resolve paths.  (If in the future that's not true, we can
be smarter.)

Test Plan:
- `make lint test`
- ran slicker on webapp over several symbols, and saw that it only spent
  any time resolving paths (or showed a progress bar) for the first
  replacement.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40687