Fix and improve syntax error handling
Summary:
When I made AST parsing lazy, I broke the handling of syntax errors,
because our excepts were no longer in the right place.  (So slicker
would crash instead of just logging an error and going on to the next
file.)  Now we'll handle them correctly, and in a less boilerplateful
way to boot.

Test Plan:
- `make lint test`
- Run slicker on itself, see error-logs instead of a crash on whichever
  file it is in `fix-includes` that `ast` for some reason doesn't like.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40709