Don't totally choke on unicode
Summary:
We were mixing unicode (from `asttokens`) and bytes (when patching and
everywhere else) unsafely, which meant that any patch we made that came
after any unicode character would cause us to patch the wrong place.
Now we use unicode and convert at the boundaries, and for `ast` which
really prefers bytes.

This is a little unfortunately factored in that it's arguably not
khodemod's responsibility to do decoding, but it was going to be much
messier to do it in slicker, because the two need to agree on whether
`Patch` takes character or byte offsets.  If we just let khodemod handle
the conversion, then everyone can just agree on character offsets which
are much more convenient.

Test Plan: `make lint test`

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40710