Add an early-exit if we know the file doesn't reference the moved symbol
Summary:
This is an optimization, to prevent having to parse and walk the AST in
cases where we can't possibly need to do anything.  Hopefully it will
speed up doing small operations on large codebases quite a bit.  We
could try to be even cleverer -- doing this before we even read the
file, for example -- but this gets most of the gains for very little
added code or complexity.

Test Plan:
- `make lint test`
- Moved an uncommonly-used name in webapp, and saw that only the
  computing paths step took more than a second or so.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40712