Prune directories when resolving paths
Summary:
Some path filters want to say "don't process this entire directory", but
in the past we traversed it anyway.  Now we don't.  To make sure the
path filter knows if it's got a file or a directory, we use a trailing
slash.  This makes resolving paths on large codebases (especially ones
with large excluded directories) much faster.

Test Plan:
- `make lint test`
- Ran `slicker.py -v` in webapp on a symbol that's not used much, saw
  that instead of 29s resolving paths it took <1s, making the whole run
  take just over a second.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40732