Only warn about import conflicts when we actually add a conflicting import.
Summary:
The old code would check if any new localnames conflicted with an
existing localname, but it would do that check too early, before we
decided if the new localname was actually needed in the file.  Thus
we'd give warnings that a new import would conflict with an existing
one even if we didn't plan on adding the new import to the file!

I moved the conflicts-check lower, to the block that decides we
actually need to add the new import.  I had to do this in two places.

Test Plan:
make check

I also ran, in webapp:
   slicker.py api/internal/permission.py permissions/api/
and got no warnings whereas before I saw things like
   ERROR:Your alias will conflict with imports in this file.
      on video_handlers.py:35 --> import permissions.util as permissions

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D40842