Don't report a conflict when moving a file but keeping the same alias.
Summary:
If we want to move profiler.py to shared/lib/profiler.py but keep the
alias the same, as we can do
    slicker.py profiler shared.lib.profiler -a profiler
The goal is to replace `import profiler` with `from shared.lib import
profiler` everywhere and leave everything else alone.

That was failing before because we add new imports before deleting old
ones, so there was a time when both imports were present at the same
time, and conflicting.

I now tell the conflict-detector to ignore the old import because
we're about to delete it.  I also make sure the deleter actually
deletes the old import.

This works well for full-file moves.  For symbol moves the old module
is sticking around, so this kind of move is not safe in general.  We
keep the TODO to allow it for the specific cases it *is* safe.

Test Plan: make check

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D40845