When matching a module-name, don't let the next symbol be `.py`.
If you see the text `foo.py` in a comment, is it referring to a file
called "foo.py", or to a function `py()` in module "foo"?  Hint: it's
the former.

We now have `_re_for_name` enforce this by not saying something is a
module-name if it's followed by `.py`.

This only matters for top-level modules, where the module-name and the
filename look the same (foo vs foo.py; as opposed to foo.bar vs
foo/bar.py where the `.`-vs-`/` saved the day before).

Auditors: benkraft

Test Plan:
make test