Add a `--use-from` flag to tell slicker to use a from-import.
Summary:
This is equivalent to `-a <basename newname>` but has better support
when moving multiple modules at once, in which case each module would
have its own newname.

I want to re-do how we specify aliases entirely, since for ADR-13 we
need a lot of flexibility there, but this is a pretty simple start.

Test Plan: make test

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: davidflanagan, carter

Differential Revision: https://phabricator.khanacademy.org/D40872