Better handle `from foo import foo`.
Summary:
When we see `from foo import foo` we ask codemod to replace both
`foo.foo` and just plain `foo` in strings and comments.  These two
patches stomp on each other (the latter tries to re-patch the patches
the former already made).  In any case, replacing just-plain-foo is
not a good idea, since the meaning of "foo" is ambiguous in this case:
is it the package or the module?

Since this only applies to things that are probably human-readable, I
think it's ok to just not replace `foo` in this case, and only replace
`foo.foo`.  Every use of a string that's not human-readable, that I
can think of, would be needing the fully qualified form anyway, so
this shouldn't hurt anything.

Test Plan: make test

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: davidflanagan, carter

Differential Revision: https://phabricator.khanacademy.org/D40916