Improve the handling of words-in-strings and words-in-comments even more.
Summary:
Files that also look like common English words can cause trouble.  For
instance `import browser` -- lots of comments refer to a browser
meaning, well, a browser, and not the browser module.

Now we are more conservative in the simple-english-word case.  This
may result in not rewriting things that we should, but hopefully it's
just in comments or some other places where it's easy for humans to
figure out what is meant.  The change to `slicker_out.py` is an
example of one such rename that we're no longer doing that we should.
(The solution may be to put "codemod" in backticks here.)

Use string.ascii_letters instead.

Back in the python 1.5.2 days, I used to import `string` all the time!
No so much now.

Test Plan:
make test

make test

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: davidflanagan, carter

Differential Revision: https://phabricator.khanacademy.org/D40915