When moving a module, keep its file-permissions.
Summary:
This was a bit awkward to thread through where it needed to go.  The
semantics are also unclear since now in theory you could have one
patch set the file-perms to X and another patch of the same file set
them to Y, and it's not clear who should win.  (I pick one arbitrarily
now.)  But I'm not sure of a better way.

I changed `Patch` from a namedtuple to a real class to support this.

Test Plan: make test

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: davidflanagan, carter

Differential Revision: https://phabricator.khanacademy.org/D40933