Fix _localname_for_localnames() when the localname is a member of a class.
Before, we assumed that if you used a locally-defined symbol in a
file, you were using a top-level symbol.  But you could also be using
a symbol defined inside the top-level symbol, such as a class var
(that is, your code uses MyClass.VAR, not just MyClass).  We weren't
handling that properly.

Now we handle the case the passed-in name has a toplevel var as a
prefix.

Auditors: benkraft

Test Plan:
make test