Ignore symlinks.  You can't move them, and we don't modify them.
Summary:
I don't know that this is strictly necessary: in practice what seems
to happen when you modify a file that's actually a symlink is that we
modify the underlying file (that is, `open(f, 'w')` opens the file
that the symlink points to, and leaves the actual link untouched).
And when we try to modify the file the second time, under its new
name, then all the codemods have already been applied via the symlink
so it's a noop, and all is good.

But I don't know if that's posix standard, and anyway it seems
confusing.  Now we notice symlinks and just skip over them.

Test Plan: make test

Reviewers: benkraft

Subscribers: davidflanagan, carter

Differential Revision: https://phabricator.khanacademy.org/D41041