Fix an edge case in implicit import handling, and a bug it was masking
This commit fixes an edge case in implicit import handling.  In
particular, the logic for deciding when to look for implicit imports in
`_localnames_from_fullnames` was incorrect.  In particular, we claimed
to (and should) be looking for the case where there is no explicit
unaliased import, but instead we looked for the case where there was no
explicit aliased import.  (Craig noticed in D41111, because the variable
name matched the correct logic, and thus not the code.)  This difference
matters when we both import a module explicitly with an alias, and use
its unaliased form (from an implicit import) -- see the new
`test_implicit_and_alias` for an example.  Additionally, it means we are
less aggressive about removing an `import foo.baz` if we are replacing
`import foo.bar` with something else -- since it isn't the best-match
import for any references we touched, we will no longer look at it when
removing.  (Thus the changes to `test_unused`, as we no longer even
tried to remove that import.)

The latter change exposed a bug in deciding whether to remove implicit
imports.  In particular, if we did `import foo.baz` and referenced
`foo.bar` later in the file, we intended to automatically remove the
import if we also found `import foo.bar`, but actually the code was only
doing so if we were also considering removing the `import foo.bar`.
Now that the above fix means we're not necessarily considering removing
that import, we need to make sure to do the right thing, and look at any
other import at the toplevel of the file.

And that logic required a further change: since we potentially recompute
some `Import` objects, we need them to have proper `__hash__` and
`__eq__` methods -- luckily they wrap the same file objects (thus the
same AST nodes when applicable) so this is just a matter of actually
implementing the methods rather than defaulting to identity.

Finally, I added another test (`test_imported_twice`) which I wrote on
the way to `test_implicit_and_alias` and figured I might as well keep
around even though it doesn't test for this bug at all.

Test plan:
`make test` passes, and fails without the changes to `slicker.py`

Reviewers: csilvers