Add support for "AUTO" aliasing, and make it the default.
Summary:
In auto-aliasing, when replacing an import we keep its alias form: a
from-import, an as-import, or a full-import.  You can still use a flag
to get the old behavior or always using a full-import for new code.

Extract import-resolution into its own function.

It's complex enough code that it deserves it!

I also removed an assumption that old_localnames is always set.

And I also added another test-case for top-level imports and AUTO.

Test Plan:
make test

make test

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: davidflanagan, carter

Differential Revision: https://phabricator.khanacademy.org/D41111