Handle the case that a localname doesn't have an import.
It seems like this happens when we're considering rewriting imports in the
file where the symbol was deleted.  We're not going to be adding an
import in that case in general -- I'm kinda confused why this function
is being called at all in this case, but maybe we haven't noticed we
don't need a new import yet by this point -- so it doesn't really
matter what we do in that case.  In any case, I wrote some tests to
make sure something reasonable happens.

Auditors: benkraft

Test plan:
Ran
   make test