When fixing, filter out names that are a proper suffix of another name.
Summary:
Consider code like this:
   fn(module_to_move.myclass, module_to_move.myclass.classvar)

If we move module_to_move to newmod, then the following names will
need to be changed:
    module_to_move.myclass -> newmod.myclass
    module_to_move.myclass.classvar -> newmod.myclass.classvar

But only the first move is actually necessary: it will fix both of
the references!  Leaving the second one in is actually a bug, since it
results in the second reference (to classvar) getting renamed twice.

Now we check for that.

Test Plan: make test

Reviewers: benkraft

Subscribers: davidflanagan, carter

Differential Revision: https://phabricator.khanacademy.org/D41340