Move cleanup suggestors to a separate file
Summary:
These are the ones that clean up whitespace, resort imports, and the
like.  But the real difference is that they have very little to do with
the rest of slicker, so they're very easy to separate out.  This is the
first of a series of commits to split up slicker.py (#6).  I did the
moves with slicker, of course, although I had to fix up a few things
manually since it doesn't like relative imports.

I also added a shortcut `make check` for `make test lint`.

Test Plan: `make check`

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D42410