Handle relative imports correctly
Summary:
Formerly, slicker didn't handle relative imports correctly.  Now it
does!  The main thing to do is to not ignore them when parsing.  Now
we parse them as usual, but set a flag on the import object saying that
it was relative, so later code can check that.  Most code doesn't need
to, though!  The module and alias are set as normal, so the vast
majority of callers don't care how the import was written -- it's just a
few edge cases.

At present we still don't attempt to handle "implicit" relative imports,
the kind that are prohibited by `absolute_import`.  We also don't
preserve relative imports with alias=AUTO -- I decided that's basically
the same as #16, although I added a skipped test.

Test Plan: `make check`

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D42662