Parse relative imports correctly
This is the first of a few commits to handle relative imports in
slicker.  The first thing we need to do is to not ignore them when
parsing.  Now, instead, we parse them as usual, but set a flag on the
import object saying that it was relative, so that if we move it
somewhere, we can preserve that.  The module and alias are set as
normal, so a lot of our code should Just Work.

At present we still don't attempt to handle "implicit" relative imports,
the kind that are prohibited by `absolute_import`.

Test plan: `make check`

Reviewers: csilvers