Fix a flaky test
This test expects slicker to make an arbitrary choice of which import's
style to follow, which it turns out to do in a way that depends on set
iteration order (of the return value of `_compute_all_imports`).  That
leads to a flaky test, although I'm not totally sure exactly what causes
the set iteration order to vary in this case.  In any case, the result
being consistent doesn't really matter in reality, so I just did some
mocking to make it consistent for the purposes of the test.  Hopefully
fixes #18.

Test plan:
`make check`, including after editing slicker.py since that seems
sometimes to trigger it

Auditors: csilvers