Add relative imports when appropriate
Summary:
D42662 set `slicker` up to correctly parse relative imports, and to
successfully pick them up and move them around.  But when rewriting
imports, we would still use absolute imports always.  Now, we replace
relative imports with relative imports when `-a AUTO` is set, and have a
new `-a RELATIVE` that forces using relative imports where it's
reasonable to do so, and falls back to from imports otherwise.

This was surprisingly much code!  Mostly because deciding whether it's
reasonble to add/keep a relative import when changing it is nontrivial,
and we need to do that kind of thing in several places.  I ended up
doing some refactoring such that instead of passing around an import
alias we intend to use in `_fix_uses_suggestor`, we simply create an
`Import` object, which may have an `alias` but also `relativity`.  Then
`_new_import_stmt` becomes a method of `Import`, which seems nice.

Fixes #16.

Test Plan: `make check`

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D42722