Split up slicker.py into a bunch of files
Summary:
This commit splits slicker.py, which has grown quite large, into a
number of different files.  What remains is not small, but is at least a
good bit more manageable.

Here are the files:
- `slicker/model.py` contains `Import`, `LocalName`, and related
  functions: that is, it contains the classes that represent slicker's
  model of the world.  I did not attempt to do any real refactoring
  (e.g. making `LocalName` a real class) at this time, but I hope this
  commit will make what needs to happen there a little clearer.
- `slicker/removal.py` contains suggestors pertaining to removing
  now-unused imports.
- `slicker/replacement.py` contains code related to replacing
  specific names in a file, namely `replace_in_file` and its helpers.
- I added a few more general utilities to `slicker/util.py`.
- Now, `slicker/slicker.py` contains only the two fix-uses suggestors,
  their helpers, and finally `make_fixes` and `main` to wrap them all
  up.  As those suggestors are really more of wrapping it all up than
  separate functionality, I kept them with `make_fixes` for now,
  although it's possible they should become yet another file.

I made a similar split of the tests, to the extent possible.  (The
majority of slicker's tests are end-to-end so this is less useful.)  I
made all the changes with `slicker` itself, using `-a RELATIVE` when
appropriate.  I had to fix up a few cases where it confused an imported
name with a variable name, and reflow some lines.

This is the remainder of what I intend to do right now to resolve #6.

Test Plan: `make check`

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D42725